Line number context and call stack in errors.
This is really a minimal support to understand where the
error could be. This is the kind of thing that is full of
details and doesn't show much general ideas. Still something
should be provided, and can be interesting to see how very
basic propagation of source-level informations work in
practice.